+++ aether/src/map_in_progress.cpp
@@ -1,6 +1,6 @@
 #include "aether/map_in_progress.hpp"
 
-CellWalls MapInProgress::get_cell_walls(int8_t x, int8_t y) const {
+CellWalls MapInProgress::get_cell_walls(int32_t x, int32_t y) const {
     y = -y; // indexing with positive y values
     return {
         horizontal_walls_[x + 1][y], // north
